Remove dependent: :destroy on tasks' taskables
This was an obvious issue since destroying any taskable would in turn destroy
tasks, which would inctroduce a recursive destruction, ultimately trying to
destroy everything all the time and then crashing.

Also, it actually makes sense not to destroy taskables when destroying tasks, as
the taskables could potentially be mapped into another task later on.

This commit is awesome!